+++ pkg/assert/assert.go
@@ -6,16 +6,17 @@
 //
 // Author: Chef (191201771@qq.com)
 
-// package assert 提供了单元测试时的断言功能，减少一些模板代码
+// Package assert 提供了单元测试时的断言功能，减少一些模板代码
 package assert
 
 import "github.com/q191201771/naza/pkg/nazareflect"
 
-// 单元测试中的 *testing.T 和 *testing.B 都满足该接口
+// TestingT 单元测试中的 *testing.T 和 *testing.B 都满足该接口
 type TestingT interface {
 	Errorf(format string, args ...interface{})
 }
 
+// TODO(chef): [refactor] 考虑将Helper方法移入 TestingT 202211
 type tHelper interface {
 	Helper()
 }
@@ -30,7 +31,7 @@ func Equal(t TestingT, expected interface{}, actual interface{}, msg ...string) 
 	return
 }
 
-// 比如有时我们需要对 error 类型不等于 nil 做断言，但是我们并不关心 error 的具体值是什么
+// IsNotNil 比如有时我们需要对 error 类型不等于 nil 做断言，但是我们并不关心 error 的具体值是什么
 func IsNotNil(t TestingT, actual interface{}, msg ...string) {
 	if h, ok := t.(tHelper); ok {
 		h.Helper()