+++ pkg/nazalog/interface.go
@@ -6,7 +6,7 @@
 //
 // Author: Chef (191201771@qq.com)
 
-// package nazalog 日志库
+// Package nazalog 日志库
 package nazalog
 
 import "errors"
@@ -25,6 +25,8 @@ import "errors"
 // * 日志文件目录不存在则自动创建
 //
 // 目前性能和标准库log相当
+//
+// TODO(chef): 异步日志
 
 var ErrLog = errors.New("naza.log:fxxk")
 
@@ -47,17 +49,29 @@ type Logger interface {
 
 	Out(level Level, calldepth int, s string)
 
-	// 断言失败后的行为由配置项Option.AssertBehavior决定
+	// Assert 断言失败后的行为由配置项Option.AssertBehavior决定
 	// 注意，expected和actual的类型必须相同，比如int(1)和int32(1)是不相等的
-	Assert(expected interface{}, actual interface{})
+	//
+	// @param expected 期望值
+	// @param actual   实际值
+	// @param extInfo  期望值和实际值不相等时打印的补充信息，如果没有，可以不填
+	//
+	Assert(expected interface{}, actual interface{}, extInfo ...string)
 
-	// flush to disk, typically
+	// Sync flush to disk, typically
+	//
 	Sync()
 
-	// 添加前缀，新生成一个Logger对象，如果老Logger也有prefix，则老Logger依然打印老prefix，新Logger打印多个prefix
+	// WithPrefix
+	//
+	// 添加前缀，新生成一个Logger对象，如果老Logger也有prefix，则老Logger依然打印老prefix，新Logger打印多个prefix。
+	//
+	// 返回的Logger对象是新的，底层的 core 是同一个
+	//
 	WithPrefix(s string) Logger
 
-	// 下面这些打印接口是为兼容标准库，让某些已使用标准库日志的代码替换到nazalog方便一些
+	// Output Print ... 下面这些打印接口是为兼容标准库，让某些已使用标准库日志的代码替换到nazalog方便一些
+	//
 	Output(calldepth int, s string) error
 	Print(v ...interface{})
 	Printf(format string, v ...interface{})
@@ -65,9 +79,21 @@ type Logger interface {
 	Fatalln(v ...interface{})
 	Panicln(v ...interface{})
 
-	// 获取配置项，注意，作用是只读，非修改配置
+	// GetOption 获取配置项
+	//
+	// 注意，作用是只读，非修改配置
+	//
 	GetOption() Option
+
+	// Init 初始化配置
+	//
+	// 注意，正常情况下，应在调用 New 函数生成Logger对象时进行配置， Init 方法提供了在已有Logger对象上配置的机会，
+	// 但是，出于性能考虑，操作logger对象内部成员时没有加锁，调用方需自行保证该函数不和其他函数并发调用，也即在使用Logger对象前（比如程序启动时)
+	//
+	Init(modOptions ...ModOption) error
 }
+
+type HookBackendOutFn func(level Level, line []byte)
 
 type Option struct {
 	Level Level `json:"level"` // 日志级别，大于等于该级别的日志才会被输出
@@ -75,9 +101,10 @@ type Option struct {
 	// 文件输出和控制台输出可同时打开
 	// 控制台输出主要用做开发时调试，打开后level字段使用彩色输出
 	Filename   string `json:"filename"`     // 输出日志文件名，如果为空，则不写日志文件。可包含路径，路径不存在时，将自动创建
-	IsToStdout bool   `json:"is_to_stdout"` // 是否以stdout输出到控制台
+	IsToStdout bool   `json:"is_to_stdout"` // 是否以stdout输出到控制台 TODO(chef): 再增加一个stderr的配置
 
-	IsRotateDaily bool `json:"is_rotate_daily"` // 日志按天翻转
+	IsRotateDaily  bool `json:"is_rotate_daily"`  // 日志按天翻转
+	IsRotateHourly bool `json:"is_rotate_hourly"` // 日志按小时翻滚，整点翻滚
 
 	ShortFileFlag       bool `json:"short_file_flag"`        // 是否在每行日志尾部添加源码文件及行号的信息
 	TimestampFlag       bool `json:"timestamp_flag"`         // 是否在每行日志首部添加时间戳的信息
@@ -85,6 +112,18 @@ type Option struct {
 	LevelFlag           bool `json:"level_flag"`             // 日志是否包含日志级别字段
 
 	AssertBehavior AssertBehavior `json:"assert_behavior"` // 断言失败时的行为
+
+	// HookBackendOutFn
+	//
+	// hook后端输出的日志内容。
+	//
+	// 业务场景：比如业务方使用了nazalog向日志文件输出日志，与之同时还想要再程序中实时获取一份日志内容。
+	//
+	// 每次回调一行日志。
+	// 获取的是全量日志。
+	// 阻塞函数。
+	// 回调结束后，内部会服用回调中日志内容的内存块。
+	HookBackendOutFn HookBackendOutFn
 }
 
 // 没有配置的属性，将按如下配置